feat(runtime): extract /share-links dispatcher domain body — ADR-0076 D11 step ③ PR-4 (#2462)
The ADR-0047 capability-token surface moves to domains/share-links.ts.
This is cloud's designed PRIMARY surface for per-env kernels
(registerShareLinkRoutes:false; host dispatcher serves after kernel swap
— the #2462 step-① re-scope finding), so behavior preservation is the
whole game here.

- DomainHandlerDeps grows getRequestKernelService: reads off the
  per-request RESOLVED kernel (this.kernel), NOT the default kernel and
  NOT the scoped-factory path — the token row and shared record must
  live in the same store as the shareLinks service's engine. The
  domain-side getEngine keeps the original try/fallback shape.
- deps also grows routeNotFound (the shared 404 envelope, 4 call sites
  in the dispatcher).
- Dispatcher keeps a thin handleShareLinks delegate; duplicated JSDoc
  removed in favor of the domain module's copy.

Verified: seam suite 30 tests (5 new: 501 absent-service, segment
boundary, UNAUTHENTICATED, public token-resolve with redaction,
ROUTE_NOT_FOUND envelope), runtime 635 green, http-conformance 41
green, dependent closure builds with DTS (--force).
